Set cycler sleep time instead of passing param through methods
The seconds to sleep between each resource is currently set by
passing the parameter from the application layer down into the
service layer. This commit changes that parameter to a class
variable that can be set in the application layer so that the param
doesn't need to be passed down into the methods of the cycler.