Better the implementation of on-demand downloads further
While the previous commits implemented better on-demand
downloading of the resources to be retrieved from the network, the
number of downloads in a cycle was limited to the number of
resources retrieved with the getter (a page at a time). This commit
introduces a download set that does not have to match the page
size of the getter's paginated results.