Add dummy resources and production profiles to project
Spring will inject beans based on the profiles given to it, so instead
of explicitly loading the dummy resource getter in the code, we
instead specify the profile. While the profile is still specified in the
code (the HtmlApplication class), later this will be refactored to
either a command line argument or a property in the config
properties file.