Load config in a @PostConstruct instead of an explicit bean
This commit makes these changes because we need to use the
system file separator in our filepaths, which is easier to put inside
of the FileUtils class. However, the file utils class has to be
constructed before it can be used in the config properties
constructor, so the functionality was moved to a post-construct
method.