Replace search box with previous search string if resumed
If the query text is changed during a cycle pause, and then the
'resume' button is hit, replace the query text with the search string
associated with the now-resumed search string. E.g. Suppose that a
cycle is paused with the search string "stream". The user then enters
a new search string of "mountain", but instead of hitting 'start',
'resume' is pressed. Then the "mountain" search string is again
replaced with the old value, i.e. "stream", and the old cycle resumes
as normal.