Fix issue with new search query not pulling correct resources
When attempting to start a cycle from a pause (i.e. after another
cycle has been started) with a different query, the different query
was not taking effect. Instead, the application would continue to
cycle through the images relevant to the initial query. This commit
fixes this issue by resetting certain resources (the resource queue
and the cycle runnable classes) when the 'start' command is invoked
again with a different query.