Fix null pointer exception thrown when null image is saved
When the 'start' command is invoked, it deletes the resources in the
resource queue and resets some of the classes for the new resou-
rce cycle. If the 'save' button is pressed before a new resource has
been assigned to the current position on the desktop, a null pointer
exception is thrown. This commit fixes this error by returning null
if the filename passed to the image save method is null or empty.