Make file utils class observable
We need a method of communicating between classes within the
model and service layers without injecting everything into everything
else. Thus we are implementing the observer pattern amongst the
classes that need to communicate. This commit makes the file utils
class observable, and the cycler runnable will observe it. The cycler
will then pass its messages to the controller, which will set the
resulting message into the gui's status field.