Add gui status message for 'next' command
Cycler runnable will now notify the cycler service of resources skip-
ped, which will then notify the gui that sets the status text. It is here
noted that the issue(s) with the odd behavior concerning the 'next'
command have not yet been resolved, so the notifications may still
be appearing at the incorrect time.