Add Bing SafeSearch 'strict' to URL
We need to make sure that the filter setting for explicit content is
set to 'strict' so that no one inadvertently ends up with explicit
images cycling on their desktop wallpaper. This commit adds the
parameter ('adult') with the 'strict' setting to the URL (hard-coded
so it can't be changed).