Use external wallpaper client to change the desktop wallpaper
After running the application on a Windows XP instance, we need
another way to set the wallpaper because the code we were using
didn't change the desktop wallpaper at all. We found a third-party
Windows wallpaper client that we will now call from the wallpaper
setter. These changes incorporate the new wallpaper client.