Add command line option to set install dir (temporarily)
Since we haven't started playing around with installing the applica-
tion yet, but we want the code to be independent of the OS (and
still be able to locate the executable directory), we will simulate
pulling the install directory from some sort of store (be it the
registry, java preferences, or whatever) by passing it in as an appl-
ication parameter. This commit implements that, as well as refact-
ors the command line object to a member variable for the html
application class.