Add quotations to paths in wallpaper setter
If a resource contains a directory that has spaces in it, then the
command will not execute properly unless the path is enclosed in
quotation marks. Furthermore, the application to be run will not be
run if the path to the wallpaper client contains spaces either. This
commit adds quotations to both of these arguments.