Replace external program with old way and convert to BMP
We suspect that XP wasn't setting the images because they weren't
BMP images. We also found out that we can't use the external
wallpaper client unless the .NET framework is installed on the OS,
and we aren't going to bother trying to get that installed whenever
a user installs this program. Therefore, we're moving back to the
old way and will convert a file to a BMP before displaying it.