Remove installation directory option and wallpaper client
The installation directory option was implemented in order to be
able to find the wallpaper client. However, the wallpaper client is
no longer going to be used since the set desktop functionality now
works in windows XP with our last change. We are therefore remov-
ing both the install directory command line option and the wallpa-
per client in this commit.