Make gui utils save dir method start from the preferred location
When the gui utils class opens up a dialogue box for the user to
select a directory in which to save resources, we want the gui's
starting folder location to be the preferred location if the preferred
location has already been set. Otherwise it will default to the temp
folder. This commit implementsthis functionality.