Add method to cycler service for setting the save directory.
The save directory has been stored in preferences only at application
startup, and it has been done so by the html application class. This
class is just an entry point for the application, however, and we
believe the functionality for setting the save directory belongs in the
cycler service. This commit adds this (not-yet used) method to the
cycler service.