Add functionality for menu item requesting save directory change
We have an option in the settings menu intended to allow the user to
change the save directory. However, the functionality has not been
implemented until now. This commit implements this functionality by
making the appropriate action call the gui controller, which calls
through to the service layer to make the change. This commit also
modifies the html application class functionality to utilize the gui
controller as well for the same purpose.