Check save dir preference before saving image to file
This commit implements functionality that checks to see if the save dir
preference is set before the image utils class saves an image that the
user has requested be saved to a file. If the preference is not set the
code execution continues, but offers a warning for why the image save
may not succeed.