Control post constructs in control panel by calling from one method
The post construct methods in the control panel were interfering
with each other as the methods in question require at least a little
order in their execution. This commit removes all of the post
construct annotations for all methods except one, which will call
the other methods and thus eliminate the errors.