Tie control panel buttons to the notification update process
Instead of just having a set amount of time that each button is
disabled after a button is clicked, this commit ties the disable
duration to the finishing of the notification update (e.g. when the
cycle has started and has notified the control panel, THEN the
buttons will be enabled).