Make logback file work
The logback file we have had in the project hasn't been working for
awhile. Instead, there has just been a default configuration applied
to the application. This commit fixes the logback file by moving it
into a place visible in the classpath (instead of inside a directory),
and also removes some unneeded stuff from the xml file.